fix history file module.
