+++ src/platform/common/logger.ts
@@ -1,0 +1,6 @@
+export interface ILogger {
+  error(errorMessage: string): void;
+  debug(debugMessage: string): void;
+  warn(warnMessage: string): void;
+  verbose(verboseMessage: string): void;
+}
