+++ src/state/globalState.ts
@@ -54,7 +54,7 @@ class GlobalState {
   public hl = true;
 
   public async load(context: vscode.ExtensionContext) {
-    this._searchHistory = new SearchHistory();
+    this._searchHistory = new SearchHistory(context);
     this._searchHistory
       .get()
       .forEach((val) =>
